pyproject.toml
[project]
name = "resources"
version = "0.1.0"
description = "Small library for managing resources"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
]
